requirements.txt
requests
python-dotenv
fastapi
uvicorn
aiofiles
pyinstaller
pywebview
pandas
openpyxl
fpdf2
python-multipart
anthropic
